move setup to TestCase
